refactor: test mojo's to shared namespace
