+++ src/main/java/ch/ivyteam/ivy/maven/test/AbstractIntegrationTestMojo.java
@@ -1,4 +1,4 @@
-package ch.ivyteam.ivy.maven;
+package ch.ivyteam.ivy.maven.test;
 
 import java.io.File;
 import java.util.Objects;
@@ -7,6 +7,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
+
+import ch.ivyteam.ivy.maven.AbstractEngineMojo;
 
 public abstract class AbstractIntegrationTestMojo extends AbstractEngineMojo
 {
