+++ src/test/java/ch/ivyteam/ivy/maven/test/TestStartEngine.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ch.ivyteam.ivy.maven;
+package ch.ivyteam.ivy.maven.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -28,9 +28,10 @@ import org.junit.Test;
 
 import com.google.common.io.Files;
 
-import ch.ivyteam.ivy.maven.AbstractIntegrationTestMojo.TestEngineLocation;
+import ch.ivyteam.ivy.maven.BaseEngineProjectMojoTest;
 import ch.ivyteam.ivy.maven.engine.EngineControl;
 import ch.ivyteam.ivy.maven.log.LogCollector;
+import ch.ivyteam.ivy.maven.test.AbstractIntegrationTestMojo.TestEngineLocation;
 
 /**
  * @since 6.1.1
